filesystem fix and some small fixes, functionality is not impacted
