--- src/Scroll.js
@@ -1,7 +1,0 @@
-const Scroll = (props) => {
-   return <div style={{overflowY: 'scroll', border:'5px solid black', height: '850px'}}>
-      {props.children}
-   </div>
-}
-
-export default Scroll
