+++ src/components/Scroll.js
@@ -1,0 +1,7 @@
+const Scroll = (props) => {
+   return <div style={{overflowY: 'scroll', border:'1px solid black', height: '850px'}}>
+      {props.children}
+   </div>
+}
+
+export default Scroll
